Move HTTP clients to their own file
